# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Web Assignment 2 API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /comment:
    get:
      tags: [Comments]
      summary: Get all comments
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      tags: [Comments]
      summary: Create a new comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreate"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Invalid input
  /comment/{id}:
    get:
      tags: [Comments]
      summary: Get comment by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Comment id (ObjectId)
      responses:
        "200":
          description: Comment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Comment not found
    put:
      tags: [Comments]
      summary: Update a comment by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Comment id (ObjectId)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentUpdate"
      responses:
        "200":
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Comment not found
    delete:
      tags: [Comments]
      summary: Delete a comment by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Comment id (ObjectId)
      responses:
        "200":
          description: Comment deleted
        "404":
          description: Comment not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ObjectId:
      type: string
      description: MongoDB ObjectId
      example: "65a1f0f3a9c2b2a1c1234567"

    # ---------- Users ----------
...
    Comment:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        post:
          $ref: "#/components/schemas/ObjectId"
        sender:
          $ref: "#/components/schemas/ObjectId"
        message:
          type: string
          example: "Nice post!"
      required: [post, sender, message]
    CommentCreate:
      type: object
      required: [post, sender, message]
      properties:
        post:
          $ref: "#/components/schemas/ObjectId"
        sender:
          $ref: "#/components/schemas/ObjectId"
        message:
          type: string
          example: "Nice post!"
    CommentUpdate:
      type: object
      properties:
        message:
          type: string
          example: "Edited comment text"